+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Moved base `services.xserver` settings to base `configuration.nix` and set `gdm` as the default display manager
+- Moved `paperwm` from `gnome` module to its own module for extra packages and config
+- Removed `config.nix` to see if it has any effect
 - Automatically import subflakes in `overlays` e.g. `overlays/paperwm`
 - Replaced `rec { }` usage with `let ... in { }`
 - Added `overlays/paperwm` subflake to track the branch `next-release` using `nix flake update` automatically